pyproject.toml
[tool.poetry]
name = "akademie-fastapi"
version = "0.1.0"
description = ""
authors = ["Jiří Lajčok <jiri.lajcok@firma.seznam.cz>", "Jiří Kupka <jiri.kupka@firma.seznam.cz>"]

[tool.poetry.dependencies]
python = "^3.7"
aiomysql = "^0.1.1"
fastapi = "^0.85.0"
httpx = "^0.23.0"
strawberry-graphql = { extras = ["fastapi"], version = "^0.133.5" }
# SQLAlchemy[asyncio] extra ensures that greenlet is installed on target platform
# https://docs.sqlalchemy.org/en/14/orm/extensions/asyncio.html#asyncio-platform-installation-notes-including-apple-m1
SQLAlchemy = { version = "^1.4.41", extras = ["asyncio"] }
uvicorn = "^0.18.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
